♻️ 그룹 정보 UI 컴포넌트 분리
